Update README.md with useful info
Add some information to the README to make it easier for someone unfamiliar with the project to understand exactly what BLEES is, how the repository is laid out, and what they need in order to work on the project. My knowledge may be somewhat out of date, so feel free to update if it is.